Configure mail and move ContactType to CommonUI module
